backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-google-genai
pydantic
groq
motor
email-validator
